Updating create_plot_Rhine_class.py
This update includes code for specific month splicing. It also provides two different plotting styles, geographical and scatter. It is built using a class set up. The majority of the user changes should happen 550 - 604. This will determine the units, the type of data being analyzed, what kind of plots are created. The titles of plots can always be changed. The scatter plots are hard coded in currently. The user will need to manually go in and change those. The units should be managed by the code, printed whatever the user chooses the units to be between lines 574-582.